+++ src/data/protocols/db/account/load-account-by-email-repository.ts
@@ -1,4 +1,4 @@
-import { AccountModel } from '../../usecases/add-account/db-add-account-protocols'
+import { AccountModel } from '../../../usecases/add-account/db-add-account-protocols'
 
 export type LoadAccountByEmailRepository = {
   loadByEmail(email: string): Promise<AccountModel | null>
