+++ src/data/protocols/db/log/log-error-repository.ts
