+++ src/data/usecases/add-account/db-add-account-protocols.ts
@@ -3,4 +3,4 @@
 export * from '../../../domain/models/account'
 export * from '../../../domain/usecases/add-account'
 export * from '../../protocols/criptography/hasher'
-export * from '../../protocols/db/add-account-repository'
+export * from '../../protocols/db/account/add-account-repository'
