+++ src/data/usecases/authentication/db-authentication-protocols.ts
@@ -2,7 +2,7 @@
 
 export * from '../../../domain/usecases/authentication'
 export * from '../../../domain/models/account'
-export * from '../../protocols/db/load-account-by-email-repository'
-export * from '../../protocols/db/update-access-token-repository'
+export * from '../../protocols/db/account/load-account-by-email-repository'
+export * from '../../protocols/db/account/update-access-token-repository'
 export * from '../../protocols/criptography/hash-comparer'
 export * from '../../protocols/criptography/encrypter'
