+++ src/presentation/controllers/login/login-controller.spec.ts
@@ -5,9 +5,9 @@ import {
   serverError,
   unauthorized,
 } from '../../helpers/http/http-helper'
-import { HttpRequest, Authentication } from './login-protocols'
-import { LoginController } from './login'
-import { Validation } from '../signup/signup-protocols'
+import { HttpRequest, Authentication } from './login-controller-protocols'
+import { LoginController } from './login-controller'
+import { Validation } from '../signup/signup-controller-protocols'
 import { AuthenticationModel } from '../../../domain/usecases/authentication'
 
 const makeFakeRequest = (): HttpRequest => ({
