+++ src/presentation/controllers/signup/signup-controller.spec.ts
@@ -2,8 +2,12 @@ import { AddAccountModel } from '../../../domain/usecases/add-account'
 import { MissingParamError } from '../../errors'
 import { badRequest, ok, serverError } from '../../helpers/http/http-helper'
 import { Validation } from '../../protocols/validation'
-import { SignUpController } from './signup'
-import { HttpRequest, AddAccount, AccountModel } from './signup-protocols'
+import { SignUpController } from './signup-controller'
+import {
+  HttpRequest,
+  AddAccount,
+  AccountModel,
+} from './signup-controller-protocols'
 
 const makeAddAccount = (): AddAccount => {
   class AddAccountStub implements AddAccount {
