+++ src/presentation/controllers/signup/signup-controller.ts
@@ -5,7 +5,7 @@ import {
   HttpRequest,
   HttpResponse,
   Validation,
-} from './signup-protocols'
+} from './signup-controller-protocols'
 
 export class SignUpController implements Controller {
   constructor(
